requirements.txt
# ChronoEdit Requirements (Merged)
# Python 3.12+
# 
# This file merges requirements.txt and requirements2.txt with conflicts resolved.

# Core ML/DL frameworks
accelerate==1.8.1
diffusers==0.35.2
peft==0.17.1
megatron-core==0.12.1
peft==0.15.2
transformer-engine==1.13.0
transformers==4.57.1
ninja
qwen-vl-utils==0.0.14
#flash-attn==2.6.3

# Data processing
numpy==1.26.4  # Using 1.26.4 for stability (2.2.6 in req2.txt may have compatibility issues)
pandas==2.2.3
pillow==11.1.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
scikit-image==0.25.2
scipy==1.15.3

# Video/Media processing
av==15.0.0
decord==0.6.0
imageio[pyav,ffmpeg]==2.37.0
imageio-ffmpeg==0.6.0
tifffile==2025.5.10
ffmpegcv

# Configuration/Hydra
hydra-core==1.3.2
omegaconf==2.3.0
pyyaml==6.0.2

# Tokenizers and text processing
sentencepiece==0.2.0
tiktoken==0.9.0
tokenizers==0.22
ftfy==6.3.1
regex==2024.11.6

# Data loading and storage
webdataset==0.2.111
zarr==2.18.3
tensorstore==0.1.45
numcodecs==0.13.1


# AWS/S3 support
boto3==1.38.31
botocore==1.38.46
s3transfer==0.13.1
fsspec==2025.5.1

# Model storage and optimization
safetensors==0.5.3
torchprofile==0.0.4

# Utilities
einops==0.8.1
attrs==25.3.0
attr==0.3.2
asciitree==0.3.3
tyro==0.9.32
tqdm==4.67.1
psutil==7.0.0
loguru==0.7.3
rich==14.0.0


# Logging/Tracking
wandb==0.21.0

# Infrastructure
fvcore==0.1.5.post20221221
iopath==0.1.10
huggingface-hub==0.35.3

# Plotting/Visualization
matplotlib==3.10.3
joblib==1.5.1

# Optional: Testing (keep if needed, remove if not running tests)
coverage==7.9.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-mock==3.14.1
pytest-random-order==1.2.0
pytest-xdist==3.8.0
ipdb==0.13.13
ipython==8.37.0

# Optional: Guardrails (keep if using content safety, remove otherwise)
better-profanity==0.7.0
nltk==3.9.1
retinaface-py==0.0.2

# Standard dependencies (required by other packages)
annotated-types==0.7.0
asttokens==3.0.0
certifi==2025.7.9
charset-normalizer==3.4.2
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
idna==3.10
importlib-metadata==8.7.0
jedi==0.19.2
jinja2==3.1.6
jmespath==1.0.1
markupsafe==3.0.2
ninja==1.11.1.4
nvidia-ml-py==12.535.133
orjson==3.11.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.8
prompt-toolkit==3.0.51
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.2
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tomli==2.2.1
tomlkit==0.13.3
traitlets==5.14.3
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.2
yacs==0.1.8
zipp==3.23.0
